// apps/frontend/packages/api/base/codegen/types/FactPackageRemark.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { FactRemarkStatus2 } from "./FactRemarkStatus2";

export interface FactPackageRemark {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type string
    */
    humanId: string;
    /**
     * @type integer
    */
    factPackageId: number;
    /**
     * @description null — замечание от ДТОиР
     * @type integer,null
    */
    cfoId?: number | null;
    /**
     * @description Обязательная привязка к форме, к которой оставлено замечание
     * @type integer
    */
    relatedFormId: number;
    /**
     * @type integer
    */
    authorId: number;
    /**
     * @type string, date-time
    */
    createdAt: string;
    /**
     * @type string
    */
    description: string;
    /**
     * @type string
    */
    requiredAction: string;
    /**
     * @type string
    */
    status: FactRemarkStatus2;
    /**
     * @type string,null, date-time
    */
    closedAt?: string | null;
    /**
     * @description Код ЦФО либо «ДТОиР»
     * @type string
    */
    issuerLabel: string;
}